Sirket_Tekillestirme/company-resolver/company-researcher/pyproject.toml
[project]
name = "company-researcher"
version = "0.0.1"
description = "Researcher agent that searches information about a company and returns it in a structured format."
authors = [
    { name = "Vadym Barda" },
    { name = "Lance Martin" }
]
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.9"
dependencies = [
    "langgraph>=0.2.52",
    "langsmith>=0.1.147",
    "langchain-community>=0.3.8",
    "tavily-python>=0.5.0",
    "langchain_anthropic>=0.3.0",
    "langchain_google_genai==2.0.10",
    "langchain-linkup",
    "bs4",
    "duckduckgo_search==6.3.5",
    "selenium==4.29.0",
    "lxml"
]

[project.optional-dependencies]
dev = ["mypy>=1.11.1", "ruff>=0.6.1"]

[build-system]
requires = ["setuptools>=73.0.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = ["agent"]
[tool.setuptools.package-dir]
"agent" = "src/agent"


[tool.setuptools.package-data]
"*" = ["py.typed"]

[tool.ruff]
lint.select = [
    "E",    # pycodestyle
    "F",    # pyflakes
    "I",    # isort
    "D",    # pydocstyle
    "D401", # First line should be in imperative mood
    "T201",
    "UP",
]
lint.ignore = [
    "UP006",
    "UP007",
    # We actually do want to import from typing_extensions
    "UP035",
    # Relax the convention by _not_ requiring documentation for every function parameter.
    "D417",
    "E501",
]
[tool.ruff.lint.per-file-ignores]
"tests/*" = ["D", "UP"]
[tool.ruff.lint.pydocstyle]
convention = "google"

[dependency-groups]
dev = [
    "langgraph-cli[inmem]>=0.1.61",
]
